+++ src/component/config.js
@@ -1,0 +1,52 @@
+/** ************************************************************************
+ *
+ * A set of configuration parameters.
+ *
+ * config.js is just a literal object that contains a set of functions. It
+ * can't be intantiated.
+ *
+ * Private Functions:
+ *  . none
+ *
+ *
+ * Public Static Methods:
+ *  . none,
+ *
+ *
+ *
+ * @namespace    -
+ * @dependencies none
+ * @exports      -
+ * @author       -
+ * @since        0.0.0
+ * @version      -
+ * ********************************************************************** */
+/* global */
+/* eslint-disable one-var, semi-style, no-underscore-dangle */
+
+
+// -- Vendor Modules
+
+
+// -- Local Modules
+
+
+// -- Local Constants
+
+
+// -- Local Variables
+
+
+// -- Public ---------------------------------------------------------------
+
+const Config = {
+
+  // the length of the component id:
+  idLength: 8,
+};
+
+
+// -- Export
+export default Config;
+
+/* eslint-enable one-var, semi-style, no-underscore-dangle */